Move `benchmark_population` from  `main.rs` to `population.rs`
As this is a helpful function in the crate I want to put it in a more appropriate place.
